+++ App/TyperF1/src/main/java/com/typerf1/typerf1/dto/grandprix/GrandPrixScoreWithJokers.java
@@ -1,4 +1,4 @@
-package com.typerf1.typerf1.dto;
+package com.typerf1.typerf1.dto.grandprix;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,19 +11,21 @@ import lombok.*;
 @Getter
 @Setter
 @NoArgsConstructor
-public class UsedJokersGP {
+public class GrandPrixScoreWithJokers {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
     String grandPrixName;
     String participantName;
     String participantSurname;
+    Long pointsSum;
     Long numberOfJokersUsed;
 
-    public UsedJokersGP(String grandPrixName, String participantName, String participantSurname, Long numberOfJokersUsed) {
+    public GrandPrixScoreWithJokers(String grandPrixName, String participantName, String participantSurname, Long pointsSum, Long numberOfJokersUsed) {
         this.grandPrixName = grandPrixName;
         this.participantName = participantName;
         this.participantSurname = participantSurname;
+        this.pointsSum = pointsSum;
         this.numberOfJokersUsed = numberOfJokersUsed;
     }
 }
