+++ App/TyperF1/src/main/java/com/typerf1/typerf1/dto/joker/JokersUsed.java
@@ -1,4 +1,4 @@
-package com.typerf1.typerf1.dto;
+package com.typerf1.typerf1.dto.joker;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,21 +11,17 @@ import lombok.*;
 @Getter
 @Setter
 @NoArgsConstructor
-public class Record {
+public class JokersUsed {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
     String participantName;
     String participantSurname;
-    String grandPrixName;
-    int year;
-    double points;
+    Long numberOfJokersUsed;
 
-    public Record(String participantName, String participantSurname, String grandPrixName, int year, double points) {
-        this.year = year;
-        this.grandPrixName = grandPrixName;
+    public JokersUsed(String participantName, String participantSurname, Long numberOfJokersUsed) {
         this.participantName = participantName;
         this.participantSurname = participantSurname;
-        this.points = points;
+        this.numberOfJokersUsed = numberOfJokersUsed;
     }
 }
