+++ App/TyperF1/src/main/java/com/typerf1/typerf1/service/ParticipantLoginDataService.java
@@ -1,14 +1,13 @@
 package com.typerf1.typerf1.service;
 
 import com.typerf1.typerf1.config.TokenGenerator;
-import com.typerf1.typerf1.dto.FullName;
+import com.typerf1.typerf1.dto.participant.FullName;
 import com.typerf1.typerf1.dto.SecurityFilterDto;
 import com.typerf1.typerf1.model.ParticipantLoginData;
 import com.typerf1.typerf1.repository.ParticipantLoginDataRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
