+++ App/TyperF1/src/main/java/com/typerf1/typerf1/service/ResultsService.java
@@ -1,8 +1,8 @@
 package com.typerf1.typerf1.service;
 
-import com.typerf1.typerf1.dto.JokersUsed;
-import com.typerf1.typerf1.dto.SeasonScore;
-import com.typerf1.typerf1.dto.SeasonScoreWithJokers;
+import com.typerf1.typerf1.dto.joker.JokersUsed;
+import com.typerf1.typerf1.dto.season.SeasonScore;
+import com.typerf1.typerf1.dto.season.SeasonScoreWithJokers;
 import com.typerf1.typerf1.repository.ResultsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
